temporary changes on the search bar flows before doing the big refactoring
