+++ .eslintrc.json
@@ -20,7 +20,8 @@
   },
   "rules": {
     "react/react-in-jsx-scope": "off",
-    "react/prop-types": "off"
+    "react/prop-types": "off",
+    "no-unused-vars": "off"
   },
   "settings": {
     "react": {
