+++ src/Pages/HomePage.css
@@ -1,33 +1,5 @@
-body {
-  margin-top: 330px;
-  padding-top: 400px;
-  width: 100vw;
-}
-
-.boxes {
+.home-container {
   display: flex;
-  justify-content: space-around;
-}
-
-.box {
-  position: relative;
-  width: 350px;
-  height: 370px;
-  padding: 30px;
-  right: 50px;
-  padding-right: 40px;
-  background-color: white;
-  color: black;
-  margin: 0 10px;
-  border-radius: 15px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  box-sizing: border-box;
-}
-
-.box h1 {
-  font-size: 4em;
-  bottom: 17rem;
-  color: #f8f697;
-  position: absolute;
-  right: 25px;
+  justify-content: center;
+  align-items: center;
 }
