+++ src/Pages/LoginPage.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useContext } from 'react'
 import axios from 'axios'
 import { Link, useNavigate } from 'react-router-dom'
-import { AuthContext } from '../context/auth.context'
+import { AuthContext } from '../context/Auth.context'
 import './LoginPage.css'
 
 const API_URL = 'http://localhost:5005'
