+++ src/components/CardCheckout.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react'
-import { CartContext } from '../context/cart.context'
+import { CartContext } from '../context/Cart.context'
 import './CardCheckout.css'
 
 const CartItem = ({ flight }) => {
