+++ src/components/Navbar.jsx
@@ -2,7 +2,7 @@
 import React, { useContext, useState } from 'react'
 import { Link } from 'react-router-dom'
 import './Navbar.css'
-import { AuthContext } from '../context/auth.context'
+import { AuthContext } from '../context/Auth.context'
 
 const Navbar = () => {
   const [isOpen, setIsOpen] = useState(false)
