+++ src/components/SearchBar.jsx
@@ -1,12 +1,46 @@
-import React, { useState } from 'react'
+import { useState } from 'react'
+
+import { useNavigate } from 'react-router-dom'
+
+// import { DataSourceContext } from '../context/DataSource.context'
+
 import './SearchBar.css'
 import DatePicker from 'react-date-picker'
 import 'react-calendar/dist/Calendar.css'
 import 'react-date-picker/dist/DatePicker.css'
 
-const SearchBar = ({ destinationInput, setDestinationInput, handleSearch }) => {
+const SearchBar = ({ setFilteredResults }) => {
+  const [flights] = useState([])
   const [selectedCategory, setSelectedCategory] = useState('All')
+  const [destinationInput, setDestinationInput] = useState('')
   const [date, setDate] = useState(new Date())
+
+  // const { initFlightsData } = useContext(DataSourceContext)
+
+  let navigateTo = useNavigate()
+
+  const handleSearch = () => {
+    navigateTo('/flights')
+    if (flights.length > 0) {
+      const filteredResults = filterData(flights)
+      console.log(filteredResults)
+      setFilteredResults(filteredResults)
+    }
+  }
+
+  const filterData = (apiData) => {
+    try {
+      return apiData.filter((data) =>
+        data.destination
+          .trim()
+          .toLowerCase()
+          .includes(destinationInput.trim().toLowerCase()),
+      )
+    } catch (error) {
+      console.error('Error filtering data:', error)
+      return []
+    }
+  }
 
   const handleDestination = (e) => {
     setDestinationInput(e.target.value)
@@ -67,7 +101,10 @@ const SearchBar = ({ destinationInput, setDestinationInput, handleSearch }) => {
           >
             <option value="All">All Categories</option>
           </select>
-          <button onClick={handleSearch} className="search-button">
+          <button
+            onClick={() => console.log('refactoring coming..')}
+            className="search-button"
+          >
             Search
           </button>
         </div>
