+++ src/main.jsx
@@ -3,16 +3,19 @@ import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
 import { BrowserRouter } from 'react-router-dom'
-import { AuthProviderWrapper } from './context/auth.context.jsx'
-import { CartContextProvider } from './context/cart.context.jsx'
+import { AuthProviderWrapper } from './context/Auth.context.jsx'
+import { CartContextProvider } from './context/Cart.context.jsx'
+import { DataSourceProvider } from './context/DataSource.context.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <BrowserRouter>
       <AuthProviderWrapper>
-        <CartContextProvider>
-          <App />
-        </CartContextProvider>
+        <DataSourceProvider>
+          <CartContextProvider>
+            <App />
+          </CartContextProvider>
+        </DataSourceProvider>
       </AuthProviderWrapper>
     </BrowserRouter>
   </React.StrictMode>,
